+++ nextgisweb/feature_layer/ogrdriver.py
@@ -2,11 +2,19 @@
 from __future__ import division, absolute_import, print_function, unicode_literals
 import collections
 
+from osgeo import ogr
+from six import ensure_str
+
+
+def test_driver_capability(name, capability):
+    driver = ogr.GetDriverByName(ensure_str(name))
+    return driver.TestCapability(capability)
+
 
 EXPORT_FORMAT_OGR = collections.OrderedDict()
 
 
-OGRDriver = collections.namedtuple(
+OGRDriverT = collections.namedtuple(
     "OGRDriver",
     [
         "name",
@@ -16,17 +24,39 @@ OGRDriver = collections.namedtuple(
         "mime",
         "single_file",
         "fid_support",
+        "lco_configurable",
+        "dsco_configurable",
     ],
 )
+
+def OGRDriver(
+    name,
+    display_name,
+    extension,
+    options=None,
+    mime=None,
+    single_file=True,
+    fid_support=False,
+    lco_configurable=None,
+    dsco_configurable=None,
+):
+    return OGRDriverT(
+        name,
+        display_name,
+        extension,
+        options,
+        mime,
+        single_file,
+        fid_support,
+        lco_configurable,
+        dsco_configurable,
+    )
 
 EXPORT_FORMAT_OGR["ESRI Shapefile"] = OGRDriver(
     "ESRI Shapefile",
     "ESRI Shapefile (*.shp)",
     "shp",
     single_file=False,
-    fid_support=False,
-    options=None,
-    mime=None,
 )
 
 EXPORT_FORMAT_OGR["GeoJSON"] = OGRDriver(
@@ -35,7 +65,6 @@ EXPORT_FORMAT_OGR["GeoJSON"] = OGRDriver(
     "geojson",
     single_file=True,
     fid_support=True,
-    options=None,
     mime="application/json",
 )
 
@@ -51,7 +80,6 @@ EXPORT_FORMAT_OGR["CSV"] = OGRDriver(
         "SEPARATOR=COMMA",
     ),
     single_file=True,
-    fid_support=False,
     mime="text/csv",
 )
 
@@ -67,7 +95,6 @@ EXPORT_FORMAT_OGR["CSV_MSEXCEL"] = OGRDriver(
         "SEPARATOR=SEMICOLON",
     ),
     single_file=True,
-    fid_support=False,
     mime="text/csv",
 )
 
@@ -76,8 +103,6 @@ EXPORT_FORMAT_OGR["DXF"] = OGRDriver(
     "AutoCAD DXF (*.dxf)",
     "dxf",
     single_file=True,
-    fid_support=False,
-    options=None,
     mime="application/dxf",
 )
 
@@ -86,9 +111,6 @@ EXPORT_FORMAT_OGR["MapInfo File (TAB)"] = OGRDriver(
     "MapInfo TAB (*.tab)",
     "tab",
     single_file=False,
-    fid_support=False,
-    options=None,
-    mime=None,
 )
 
 EXPORT_FORMAT_OGR["MapInfo File (MIF/MID)"] = OGRDriver(
@@ -96,9 +118,6 @@ EXPORT_FORMAT_OGR["MapInfo File (MIF/MID)"] = OGRDriver(
     "MapInfo MIF/MID (*.mif/*.mid)",
     "mif",
     single_file=False,
-    fid_support=False,
-    options=None,
-    mime=None,
 )
 
 EXPORT_FORMAT_OGR["GPKG"] = OGRDriver(
@@ -107,8 +126,16 @@ EXPORT_FORMAT_OGR["GPKG"] = OGRDriver(
     "gpkg",
     single_file=True,
     fid_support=True,
-    options=None,
     mime="application/geopackage+vnd.sqlite3",
+)
+
+EXPORT_FORMAT_OGR["SXF"] = OGRDriver(
+    "SXF",
+    "Storage and eXchange Format (*.sxf)",
+    "sxf",
+    single_file=False,
+    options=("SXF_NEW_BEHAVIOR=YES",),
+    dsco_configurable=("SXF_MAP_SCALE:1000000", "SXF_MAP_NAME", "SXF_SHEET_KEY"),
 )
 
 
@@ -117,6 +144,9 @@ OGR_DRIVER_NAME_2_EXPORT_FORMATS = [
         "name": format_id,
         "display_name": format.display_name,
         "single_file": format.single_file,
+        "lco_configurable": format.lco_configurable,
+        "dsco_configurable": format.dsco_configurable,
     }
     for format_id, format in EXPORT_FORMAT_OGR.items()
+    if test_driver_capability(format.name, ogr.ODrCCreateDataSource)
 ]